+++ lib/dartapi_core.dart
@@ -1,9 +1,10 @@
 library;
 
+export 'src/core/api_exception.dart';
+export 'src/core/api_methods.dart';
 export 'src/core/api_route.dart';
 export 'src/core/base_controller.dart';
-export 'src/core/serializable.dart';
 export 'src/core/logger.dart';
-export 'src/core/api_methods.dart';
+export 'src/core/serializable.dart';
 
 export 'src/utils/utils.dart';